streamlit_app/requirements.txt
google-cloud-bigquery
db-dtypes
streamlit
pandas
altair
seaborn